Made changes to docker-compose file to add PostgresDB container only if it doesn't exist and fixed Signin/Signup pages to interact with REST API, persisting Postgres data even after running "docker-compose up --build" command.
have made changes to the docker-compose file to add the PostgresDB container only if it doesn't already exist. Additionally, the container will create all the required tables without requiring external creation.
Whenever we run "docker-compose up --build," it will recreate only the ReactApp and JavaAPIApp, but not the Postgres container if it already exists. This means that the data within the Postgres container will persist, even after running the command.
In addition , fixed the Signin and Signup pages. Both of these pages are now able to interact with the REST API application. This means that users can now create new accounts and sign in to existing ones, and their information will be stored and retrieved using the REST API